refactor : 디렉토리 세분화해서 변경
