+++ backend/src/main/java/com/project/Tamago/controller/UserController.java
@@ -1,0 +1,14 @@
+package com.project.Tamago.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import lombok.RequiredArgsConstructor;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/user")
+public class UserController {
+
+
+}
