+++ backend/src/main/java/com/project/Tamago/dto/requestDto/LoginReqDto.java
@@ -1,4 +1,4 @@
-package com.project.Tamago.dto;
+package com.project.Tamago.dto.requestDto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
