+++ backend/src/main/java/com/project/Tamago/dto/responseDto/ProfileResDto.java
@@ -1,0 +1,4 @@
+package com.project.Tamago.dto;
+
+public class ProfileResDto {
+}
