+++ backend/src/main/java/com/project/Tamago/mapper/UserMapper.java
@@ -1,0 +1,2 @@
+package com.project.Tamago.mapper;public class UserMapper {
+}
