+++ backend/src/main/java/com/project/Tamago/service/UserService.java
@@ -1,0 +1,2 @@
+package com.project.Tamago.service;public class UserService {
+}
